% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{close.RethinkDB_connection}
\alias{close.RethinkDB_connection}
\title{Close RethinkDB connection}
\usage{
\method{close}{RethinkDB_connection}(con, ...)
}
\arguments{
\item{con}{Connection to close.}

\item{...}{Ignored.}
}
\description{
Closes connection and stops all associated callbacks and/or sync cursor.
}

