% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cursor.R
\name{close.RethinkDB_cursor}
\alias{close.RethinkDB_cursor}
\title{Close cursor}
\usage{
\method{close}{RethinkDB_cursor}(con, ...)
}
\arguments{
\item{con}{Cursor to close.}

\item{...}{Ignored.}
}
\description{
Closes a given cursor and stops its associated query.
Should be called on the current cursor before a new sync query is invoked on the same connection.
}

