% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cursor.R
\name{cursorToList}
\alias{cursorToList}
\title{Convert cursor into a list}
\usage{
cursorToList(x, maxResults = 10000)
}
\arguments{
\item{x}{RethinkDB cursor or any object.}

\item{maxResults}{Number of results sufficient to stop pulling from cursor.}
}
\value{
A list of elements pulled from \code{x} if it is a cursor, \code{x} otherwise.
}
\description{
Converts cursor into a list.
For convenience, when given anything other than cursor returns this object unchanged; this way can be used to wrap the result of \code{$run}, so that it is never a cursor.
}
\note{
The lenght of a list may be larger than \code{maxResults} because RethinkDB transmits results in batches.
}
\author{
Miron B. Kursa
}

