% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{isOpened}
\alias{isOpened}
\title{Check if connection is opened}
\usage{
isOpened(x)
}
\arguments{
\item{x}{Connection to to check.}
}
\value{
\code{TRUE} if connection is opened and can be used for queries, \code{FALSE} otherwise.
}
\description{
Check whether a given connection is opened.
Closed connections cannot be used, and will throw errors on such an attempt; their associated callbacks and/or sync cursor are dead and won't fire/produce any more data.
}
\author{
Miron B. Kursa
}
