% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{openConnection}
\alias{openConnection}
\title{Open connection to a RethinkDB server}
\usage{
openConnection(host = "localhost", port = 28015, authKey = NULL,
  v = "V0_4")
}
\arguments{
\item{host}{Host to connect to.}

\item{port}{Port to connect to.}

\item{authKey}{Authentication key. Not supported yet.}

\item{v}{Protocol version; \code{"V0_3"} and \code{"V0_4"} supported, the default should be used.}
}
\value{
Object of a class \code{RethinkDB_connection}, which can be passed to \code{r()$...$run} and \code{r()$...$runAsync} functions.
}
\description{
Opens connection to a given RethinkDB server.
}
