% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{print.RethinkDB_connection}
\alias{print.RethinkDB_connection}
\title{Print RethinkDB connection}
\usage{
\method{print}{RethinkDB_connection}(x, ...)
}
\arguments{
\item{x}{Connection to print.}

\item{...}{Ignored.}
}
\description{
Prints a RethinkDB connection details, including a number of pending callbacks.
}
\note{
Never blocks.
}
