% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cursor.R
\name{print.RethinkDB_cursor}
\alias{print.RethinkDB_cursor}
\title{Print cursor}
\usage{
\method{print}{RethinkDB_cursor}(x, ...)
}
\arguments{
\item{x}{Cursor to print.}

\item{...}{Ignored.}
}
\description{
Prints a given cursor's status.
}
\note{
Never blocks; also checks whether the underlying connection is alive.
}

